+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Tue Mar 29 14:49:52 GMT+02:00 2022
-gradle.version=7.2
+#Tue Dec 27 20:17:43 IST 2022
+gradle.version=7.4
