+++ app/src/main/AndroidManifest.xml
@@ -17,6 +17,9 @@
         android:usesCleartextTraffic="true">
 
         <activity
+            android:name=".Activities.welcomepage"
+            android:exported="true" />
+        <activity
             android:name=".Activities.acivity.OpenEnded"
             android:configChanges="orientation|screenSize"
             android:exported="false" />
@@ -27,8 +30,8 @@
         <activity
             android:name=".Activities.QuizActivity"
             android:configChanges="orientation|screenSize"
-            android:windowSoftInputMode="stateAlwaysHidden|adjustResize"
-            android:exported="false" />
+            android:exported="false"
+            android:windowSoftInputMode="stateAlwaysHidden|adjustResize" />
         <activity
             android:name=".Activities.acivity.TruefalseActivity"
             android:configChanges="orientation|screenSize"
@@ -62,7 +65,8 @@
             android:configChanges="orientation|screenSize"
             android:theme="@style/Theme.AppCompat.DayNight.NoActionBar"
             android:windowSoftInputMode="stateAlwaysHidden|adjustResize" />
-        <activity android:name=".Activities.Register"
+        <activity
+            android:name=".Activities.Register"
             android:windowSoftInputMode="stateAlwaysHidden|adjustResize">
 
             <!-- <intent-filter> -->
@@ -75,6 +79,7 @@
         <activity
             android:name=".Activities.Splash"
             android:theme="@style/Theme.AppCompat.DayNight.NoActionBar">
+            >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -84,9 +89,9 @@
         <activity
             android:name=".Activities.Mcqs_activity"
             android:exported="true" />
-        <activity android:name=".Activities.Login"
-            android:windowSoftInputMode="stateAlwaysHidden|adjustResize"
-            >
+        <activity
+            android:name=".Activities.Login"
+            android:windowSoftInputMode="stateAlwaysHidden|adjustResize">
 
             <!-- <intent-filter> -->
             <!-- <action android:name="android.intent.action.MAIN" /> -->
