+++ app/src/main/res/drawable-mdpi/custom_button.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+<gradient android:startColor="#2AF598" android:endColor="#009EFD" android:type="linear"/>
+    <corners android:radius="30dp"/>
+</shape>
