+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Tue Mar 29 04:49:46 GMT-08:00 2022
-sdk.dir=C\:\\Users\\aliza\\AppData\\Local\\Android\\Sdk
+#Tue Dec 27 20:17:05 IST 2022
+sdk.dir=C\:\\Users\\sakth\\AppData\\Local\\Android\\Sdk
